Take Into Consideration Spree Store Credit
When a PayPal express payment is performed, we should take into
consideration the store credits that will be consumed during the
following checkout. Pass the order total with store credits over just
the total. Along with this, pass the store credit into the items array
with a negative number to come out with a correct amount to charge the
user.

Assembla:
https://www.assembla.com/spaces/hollar-development/tickets/new_cardwall#ticket:60